#89 Change file structure to a new datastreams-folder

+++ backend/datastreams/openface/openface.py
@@ -7,6 +7,7 @@ class OpenFaceInstance:
 
     process = None
     out_filename = "data"
+    out_dir = ".\datastreams\openface\processed"
     device_id = 0
 
     file = None
@@ -16,18 +17,26 @@ class OpenFaceInstance:
 
     def __init__(self, out_dir, out_filename, device_id):
         self.out_filename = out_filename
+        self.out_dir = out_dir
         self.device_id = device_id
 
     def __init__(self):
         pass
 
     def start(self):
-        self.process = Popen("\".\openface2\FeatureExtraction.exe\"" +
-                             " -device " + str(self.device_id) +
-                             " -of " + "\"" + self.out_filename + "\"")
+        try:
+            self.process = Popen("\".\datastreams\openface\\bin\FeatureExtraction.exe\"" +
+                                 " -device " + str(self.device_id) +
+                                 " -of " + "\"" + self.out_filename + "\"" +
+                                 " -out_dir " + self.out_dir)
+        except FileNotFoundError:
+            self.process = None
+            print(
+                "Couldn't find the OpenFace binaries, make sure they are installed correctly.")
 
     def terminate(self):
-        self.process.terminate()
+        if self.process:
+            self.process.terminate()
 
     def startDataRead(self):
         self.file = open("./processed/data.csv", newline="")
